refactor(agent): complete Action Loop responsibility split
