+++ framework/core/src/python/kungfu/agent/_action_loop_transport.py
@@ -2,31 +2,39 @@
 
 from __future__ import annotations
 
-import argparse
-import json
-import sys
 from collections.abc import Mapping
 from pathlib import Path
 from typing import Any, Callable
 
 
-def dispatch(runtime_dir: str | Path, operation: str, payload: Any) -> Any:
-    from kungfu.agent import action_loop
+def validate_native_authority(
+    runtime_dir: str | Path,
+    payload: Any,
+    inspector: Callable[[str | Path, Mapping[str, Any]], dict[str, Any]],
+) -> dict[str, Any] | None:
+    if not isinstance(payload, Mapping):
+        return None
+    expected = payload.get("nativeAuthority")
+    envelope = payload.get("envelope")
+    if expected is None and isinstance(envelope, Mapping):
+        expected = envelope.get("nativeAuthority")
+    if not isinstance(expected, Mapping):
+        return None
+    authority = inspector(runtime_dir, expected)
+    return authority if authority.get("status") != "current" else None
 
+
+def authority_gate(runtime_dir, operation, payload, inspector) -> tuple[bool, Any]:
     if operation == "authority-inspect":
         expected = payload if isinstance(payload, Mapping) else None
-        return action_loop.inspect_native_authority(runtime_dir, expected)
-    if isinstance(payload, Mapping):
-        expected_authority = payload.get("nativeAuthority")
-        envelope = payload.get("envelope")
-        if expected_authority is None and isinstance(envelope, Mapping):
-            expected_authority = envelope.get("nativeAuthority")
-        if isinstance(expected_authority, Mapping):
-            authority = action_loop.inspect_native_authority(
-                runtime_dir, expected_authority
-            )
-            if authority.get("status") != "current":
-                return authority
+        return True, inspector(runtime_dir, expected)
+    authority = validate_native_authority(runtime_dir, payload, inspector)
+    return authority is not None, authority
+
+
+def operation_handler(operation: str):
+    from kungfu.agent import action_loop
+
     operations = {
         "work-profile-bind": action_loop.bind_work_profile,
         "episode-resume-or-begin": action_loop.resume_or_begin_episode,
@@ -42,27 +50,4 @@ def dispatch(runtime_dir: str | Path, operation: str, payload: Any) -> Any:
     handler = operations.get(operation)
     if handler is None:
         raise ValueError(f"unsupported Action Loop adapter operation: {operation}")
-    return handler(runtime_dir, payload)
-
-
-def main(
-    dispatcher: Callable[[str | Path, str, Any], Any],
-    argv: list[str] | None = None,
-) -> int:
-    parser = argparse.ArgumentParser()
-    parser.add_argument("--runtime-dir", required=True)
-    parser.add_argument("operation")
-    args = parser.parse_args(argv)
-    try:
-        payload = json.load(sys.stdin)
-        result = dispatcher(args.runtime_dir, args.operation, payload)
-    except Exception as error:
-        result = {
-            "status": "denied",
-            "code": "adapter-error",
-            "message": str(error),
-            "writeOccurred": False,
-        }
-    json.dump(result, sys.stdout, ensure_ascii=False, separators=(",", ":"))
-    sys.stdout.write("\n")
-    return 0
+    return handler
